+++ tests/base/packageC/module_c2.py
@@ -1,4 +1,4 @@
-from src.decorator import load_config
+from autoload.decorator import load_config
 from tests.test_module import TestModule
